Update monitor.js
This additional 175 lines of code makes it so the Fitness Monitor 1. Takes in data and logs it correctly to the right date. 2. Displays the current month & year at the calendar header and allows the user to access data for previous and coming months with a click of the right and left arrow buttons in their respective places on the header, the user is also able to click the month on the header to pick exactly what month & year to go to.  3. Displays a Progress bar and updates it as the user adds data. 4. prompts the user to either update, add or delete data by clicking on a specific day on the calendar although the edit button functionality is not yet complete.